+++ sphaira/source/utils/devoptab_nsp.cpp
@@ -1,5 +1,6 @@
 
 #include "utils/devoptab.hpp"
+#include "utils/devoptab_common.hpp"
 #include "defines.hpp"
 #include "log.hpp"
 
@@ -17,7 +18,7 @@ namespace sphaira::devoptab {
 namespace {
 
 struct Device {
-    std::unique_ptr<yati::source::Base> source;
+    std::unique_ptr<common::BufferedData> source;
     yati::container::Collections collections;
 };
 
@@ -43,7 +44,7 @@ int devoptab_open(struct _reent *r, void *fileStruct, const char *_path, int fla
     std::memset(file, 0, sizeof(*file));
 
     char path[FS_MAX_PATH];
-    if (!fix_path(_path, path)) {
+    if (!common::fix_path(_path, path)) {
         return set_errno(r, ENOENT);
     }
 
@@ -111,7 +112,7 @@ DIR_ITER* devoptab_diropen(struct _reent *r, DIR_ITER *dirState, const char *_pa
     std::memset(dir, 0, sizeof(*dir));
 
     char path[FS_MAX_PATH];
-    if (!fix_path(_path, path)) {
+    if (!common::fix_path(_path, path)) {
         set_errno(r, ENOENT);
         return NULL;
     }
@@ -166,7 +167,7 @@ int devoptab_lstat(struct _reent *r, const char *_path, struct stat *st) {
     log_write("[\t\tDEV] lstat\n");
 
     char path[FS_MAX_PATH];
-    if (!fix_path(_path, path)) {
+    if (!common::fix_path(_path, path)) {
         return set_errno(r, ENOENT);
     }
 
@@ -236,7 +237,12 @@ Result MountNsp(fs::Fs* fs, const fs::FsPath& path, fs::FsPath& out_path) {
     }
 
     auto source = std::make_unique<yati::source::File>(fs, path);
-    yati::container::Nsp nsp{source.get()};
+
+    s64 size;
+    R_TRY(source->GetSize(&size));
+    auto buffered = std::make_unique<common::BufferedData>(std::move(source), size);
+
+    yati::container::Nsp nsp{buffered.get()};
     yati::container::Collections collections;
     R_TRY(nsp.GetCollections(collections));
 
@@ -245,7 +251,7 @@ Result MountNsp(fs::Fs* fs, const fs::FsPath& path, fs::FsPath& out_path) {
     entry.devoptab = DEVOPTAB;
     entry.devoptab.name = entry.name;
     entry.devoptab.deviceData = &entry.device;
-    entry.device.source = std::move(source);
+    entry.device.source = std::move(buffered);
     entry.device.collections = collections;
     std::snprintf(entry.name, sizeof(entry.name), "nsp_%u", g_mount_idx);
     std::snprintf(entry.mount, sizeof(entry.mount), "nsp_%u:/", g_mount_idx);
